Create duke package to contain all. Create duke.storage, duke.task and duke.commands for those classes respectively
