+++ src/main/java/duke/Duke.java
@@ -1,4 +1,7 @@
-import java.util.Optional;
+package duke;
+
+import duke.commands.Command;
+import duke.storage.Storage;
 
 public class Duke {
