+++ src/main/java/duke/DukeExceptions.java
@@ -1,5 +1,7 @@
+package duke;
+
 public class DukeExceptions extends Exception{
-    DukeExceptions(String msg) {
+    public DukeExceptions(String msg) {
         super(msg);
     }
 }
