+++ src/main/java/duke/commands/Command.java
@@ -1,3 +1,8 @@
+package duke.commands;
+
+import duke.Ui;
+import duke.storage.Storage;
+
 public abstract class Command {
     public abstract boolean execute(Ui ui, Storage storage);
 }
