+++ src/main/java/duke/commands/DeadlineCommand.java
@@ -1,3 +1,9 @@
+package duke.commands;
+
+import duke.Ui;
+import duke.storage.Storage;
+import duke.tasks.Deadline;
+
 public class DeadlineCommand extends Command{
     Deadline deadline;
     public DeadlineCommand(Deadline deadline) {
