+++ src/main/java/duke/commands/DeleteCommand.java
@@ -1,21 +1,23 @@
-public class DoneCommand extends Command{
+package duke.commands;
+
+import duke.Ui;
+import duke.storage.Storage;
+
+public class DeleteCommand extends Command{
     int index;
 
-    public DoneCommand (int input) {
+    public DeleteCommand (int input) {
         this.index = input - 1;
     }
-
 
     @Override
     public boolean execute(Ui ui, Storage storage) {
         try {
-            storage.markAsFinished(index);
-            ui.print("Well done! I have marked the following as finished: \n" + storage.getTask(index));
+            ui.print("Okay! I have deleted this task from your list: \n" + storage.getTask(index));
+            storage.deleteFromList(index);
             storage.save();
         } catch (IndexOutOfBoundsException e){
             ui.print("Oops, the list is not that big!");
-        } catch (DukeExceptions e) {
-            ui.printException(e);
         }
         return false;
     }
