+++ src/main/java/duke/commands/ListCommand.java
@@ -1,4 +1,9 @@
-public class ListCommand extends Command{
+package duke.commands;
+
+import duke.Ui;
+import duke.storage.Storage;
+
+public class ListCommand extends Command {
 
     @Override
     public boolean execute(Ui ui, Storage storage) {
