+++ src/main/java/duke/tasks/ToDo.java
@@ -1,3 +1,5 @@
+package duke.tasks;
+
 public class ToDo extends Task{
 
     private ToDo(String description) {
